requirements.txt
streamlit==1.45.0
tensorflow==2.19.0
numpy==1.26.4
pandas==2.2.3
requests>=2.31.0
Pillow>=10.0.0
matplotlib>=3.8.0
pytest
pytest-mock